refactor: extract streaming eviction controller
Move eviction state, LRU selection, slot-pressure eviction, and
per-frame counter management out of GaussianStreamingSystem into
StreamingEvictionController. Eviction config loading moved out of
StreamingUploadPipeline into the controller. All counter/timestamp
mutation sites now go through controller methods. Semantics preserved:
requested-residency increments total only, non-primary LRU conditionally
increments visible, ensure_atlas_slot_available ignores requesting_asset_id.
